fix .net warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6727 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/SpawnPosition.cxx
@@ -45,7 +45,7 @@ SpawnPosition::SpawnPosition(int playerId, bool onGroundOnly, bool notNearEdges)
   } else {
     const float tankHeight = BZDB.eval(StateDatabase::BZDB_TANKHEIGHT);
     const float tankRadius = BZDB.eval(StateDatabase::BZDB_TANKRADIUS);
-    safeSWRadius = (BZDB.eval(StateDatabase::BZDB_SHOCKOUTRADIUS) + BZDB.eval(StateDatabase::BZDB_TANKRADIUS)) * 1.5;
+    safeSWRadius = (float)((BZDB.eval(StateDatabase::BZDB_SHOCKOUTRADIUS) + BZDB.eval(StateDatabase::BZDB_TANKRADIUS)) * 1.5);
     safeDistance = tankRadius * 20; // FIXME: is this a good value?
     const float size = BZDB.eval(StateDatabase::BZDB_WORLDSIZE);
     const float maxWorldHeight = world->getMaxWorldHeight();
